Create TriangleSort.lua
Triangle sort is named such because it uses a turtle and three chests in a -- you guessed it -- triangle. The turtle should have a modem attached or a solid block modem next to it that it can connect to.
The chests should either be inputs from other sorting turtles or outputs to machinery. I'm working on adding more complicated machinery and on demand crafting.